Use the smallest needed datatype for ultrasonic

+++ lib/drv_ultrasonic/ultrasonic.cpp
@@ -4,14 +4,14 @@
 // this constant won't change. It's the pin number of the sensor's output:
 #define PING_PIN 9
 
-uint64_t microsecondsToCentimeters(int64_t microseconds) {
+uint16_t microsecondsToCentimeters(uint32_t microseconds) {
     // The speed of sound is 340 m/s or 29 microseconds per centimeter.
     // The ping travels out and back, so to find the distance of the object we
     // take half of the distance travelled.
     return microseconds / 29 / 2;
 }
 
-uint64_t ultrasonic::distance() {
+uint16_t ultrasonic::distance() {
     // The PING))) is triggered by a HIGH pulse of 2 or more microseconds.
     // Give a short LOW pulse beforehand to ensure a clean HIGH pulse:
     pinMode(PING_PIN, OUTPUT);
@@ -28,7 +28,7 @@ uint64_t ultrasonic::distance() {
 
     // Do not disable interrupts, as it fucks with the Servo.h library
     // noInterrupts();
-    uint64_t duration = pulseIn(PING_PIN, HIGH);
+    uint32_t duration = pulseIn(PING_PIN, HIGH);
     // interrupts();
 
     return microsecondsToCentimeters(duration);
